+++ TMenu.h
@@ -11,7 +11,7 @@ void MENU_Motor(void);
 
 // Configuration getters
 BOOL MENU_isConfigured(void);
-const WORD *MENU_GetTMPThresholds(void);
+const BYTE *MENU_GetTMPThresholds(void);
 BYTE MENU_GetSamplingTime(void);
 const BYTE *MENU_GetInitialTimeString(void);
